unit_test: Move some interp test helpers to a separate header for easier reuse.
