+++ unit_tests/interp/meson.build
@@ -2,3 +2,5 @@ test_interp_srcs = files([
   'tests_main.cc',
   'test_interp_basics.cc',
   ])
+
+test_interp_inc = include_directories('.')
